refactor(import): split TimezonePicker into picker, item, and catalog
TimezonePicker.tsx is down from ~300 to ~90 lines. The catalog logic
(useTimezoneCatalog + IANA-zone helpers) moves to timezoneCatalog.ts,
and each CommandItem row is now a TimezoneItem component.
